refactor(sign in-up pages): organize Sign in-up pages component
Add a "Shared Components" folder to the project to organize and store reusable components that can
be used across the Sign In/Up pages, to improve code structure and maintainability
