+++ src/components/__tests__/ButtonAuth.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import ButtonAuth from "../ButtonAuth";
+import ButtonAuth from "@/components/shared/ButtonAuth";
 
 it("renders correctly", () => {
     const tree = renderer.create(<ButtonAuth />).toJSON();
