+++ src/components/__tests__/FormFooter.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import FormFooter from "@/components/FormFooter";
+import FormFooter from "@/components/shared/FormFooter";
 
 it("renders correctly", () => {
     const tree = renderer.create(<FormFooter />).toJSON();
