+++ src/components/__tests__/InputForm.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import InputForm from "../InputForm";
+import InputForm from "@/components/shared/InputForm";
 
 it("renders correctly", () => {
     const tree = renderer.create(<InputForm />).toJSON();
