+++ src/components/__tests__/OrDivider.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import OrDivider from "../OrDivider";
+import OrDivider from "@/components/shared/OrDivider";
 
 it("renders correctly", () => {
     const tree = renderer.create(<OrDivider />).toJSON();
