+++ src/components/__tests__/SignButton.test.jsx
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import SignButton from "../SignButton";
+import SignButton from "@/components/shared/SignButton";
 
 it("renders correctly", () => {
     const tree = renderer.create(<SignButton />).toJSON();
