+++ src/components/shared/SittingImg.jsx
